move common method
